backend/requirements.txt
# 商探AI Backend Dependencies
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pgvector==0.3.6

# Cache
redis==5.2.1

# Auth
pyjwt==2.10.1
bcrypt==4.2.1
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.28.1

# Validation
pydantic==2.10.4
pydantic-settings==2.7.1

# Resilience
tenacity==9.0.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.20

# Task Queue
celery[redis]==5.4.0

# Encryption
cryptography==44.0.0
